experimenting with react-image for loading placeholders while image loads
With this method you can load a 20px x 20px thumb of the full image as a placeholder, stretch it to the same size as the image, and fade in the final image once it's loaded (progressive image loading technique used of Medium). 
This will keep the UX nicer by fixing the issue that images may load slowly while scrolling due to them being lazy loaded. Using this method the user will see a placeholder (blurred version of the original) while the image is loading.